Convert to inttypes.h conversion specifiers
Change the printf conversion specifiers to use the portable
macros from inittypes.h.  For instance, instead of guessing %lu for
a uint32_t, use the PRIu32 macro.

Also change the length modifier for size_t types to "z", which is
an explicit printf length modifier for size_t and ssize_t arguments.

Remove uses of The %m conversion specifier.  While it is convenient,
it is explicitly listed as a Glibc extension and therefore less
portable.  In subprocess.c, the %m was already redundant with the
error string appened by log_err(), so they were simply removed.  The
one final instance of %m in popen2.c was converted to an explicit
%s and strerr(errno).

Closes #394

+++ doc/man3/topen.c
@@ -1,3 +1,4 @@
+#include <inttypes.h>
 #include <flux/core.h>
 #include "src/common/libutil/log.h"
 
@@ -10,7 +11,7 @@ int main (int argc, char **argv)
         log_err_exit ("flux_open");
     if (flux_get_rank (h, &rank) < 0)
         log_err_exit ("flux_get_rank");
-    printf ("My rank is %d\n", (int)rank);
+    printf ("My rank is %"PRIu32"\n", rank);
     flux_close (h);
     return (0);
 }
